Remove `crypto` package and refactor to `cryptography`
Removed the `crypto` package from the internal/algorithm directory. This was done because the implementations were leveraging an external security library. To improve code quality and control, they have been rewritten, restructured into the new `cryptography` package, and more coherent and comprehensive tests have been added.

Also, direct handling of error was implemented in the `compression.go` file, and the dependency on `helpers` package was removed, increasing readability and efficiency of the code.