Add key generation methods to cryptography algorithm
This update introduces new cryptographic key generation methods which use a map to return specific key calculations, based on an input key. It adds extra checks for input key length and increases test coverage for the new feature. The `ConvertToKeys` function which was previously in the `cryptography.go` is now in the new file `keys.go` which is responsible for key transformations. Code for the key generation is now housed in `generate_keys.go` where a new file is created as a result of its execution, `generated_file.go` containing parts of the cryptographic solution for later use. This change aims to increase readability and maintainability of the project as now `cryptography.go` is fully dedicated for crypto-based algorithms, whereas routines to generate cryptographic keys are separated.